projects/doc-rag/requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
faiss-cpu
transformers
torch
python-multipart
pydantic
tqdm